msm: pm: debug counters to DDR
Today the PC debug counters are placed in IMEM. There is
a need for more memory to save these counters for extra
CPUs in newer platforms. As IMEM is scarce resource, move
these per CPU counters to non cacheable area in DDR.
Also save the physical start address in IMEM for easy
access to location of counters while analyzing dumps.

Signed-off-by: Venkat Devarasetty <vdevaras@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>