msm-core: Read frequencies from cpufreq driver
OPP interface provides many more frequencies than those available for
the cpus. We generate power numbers for more frequencies than we need,
and provide these to the scheduler. While this works, it is not required
to save information of unavailable frequencies.

Read the available frequencies from cpufreq driver and calculate power
numbers only for these frequencies.

Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>