msm-core: Fix NULL pointer issue while repopulating stats
There is a small window between when cpu is brought online and when
CPUFREQ_CREATE_POLICY notification is received. If we end up here within
this window for that cpu, we might try to populate stats even before
cpufreq policy is updated. This could result in a NULL pointer error.

Wait for frequency and ptable update before repopulating the stats.

Signed-off-by: Archana Sathyakumar <asathyak@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>